Revert "*Move to getgradient"
This reverts commit 6cfba8030a54a70889d91b67c3eca4e49908979d.

+++ src/volume/VoxelGradient.java
@@ -10,18 +10,18 @@ package volume;
  */
 public class VoxelGradient {
 
-    public double x, y, z;
-    public double mag;
+    public float x, y, z;
+    public float mag;
     
     public VoxelGradient() {
         x = y = z = mag = 0.0f;
     }
     
-    public VoxelGradient(double gx, double gy, double gz) {
+    public VoxelGradient(float gx, float gy, float gz) {
         x = gx;
         y = gy;
         z = gz;
-        mag = (double) Math.sqrt(x*x + y*y + z*z);
+        mag = (float) Math.sqrt(x*x + y*y + z*z);
     }
     
 }
